pcap: refactor: rename capability classes
