--- src/main/java/it/smartphonecombo/uecapabilityparser/model/pcap/OsmoCoreLog.kt
@@ -1,6 +1,0 @@
-package it.smartphonecombo.uecapabilityparser.model.pcap
-
-class OsmoCoreLog(
-    val text: String,
-    val isNr: Boolean,
-)
