+++ src/main/java/it/smartphonecombo/uecapabilityparser/model/pcap/capability/CaCombosSupported.kt
@@ -1,0 +1,7 @@
+package it.smartphonecombo.uecapabilityparser.model.pcap.capability
+
+// qcom 0xB0CD, 0xB826
+class CaCombosSupported(
+    val text: String,
+    val isNr: Boolean,
+)
